Add task to state of CourseCreate page on select from drop down
Next steps are to pass the course's tasks to a component that will allow for the dragging and dropping of task order. Will also need to populate this edit task page with props in the form of the current course in the redux store. A lot of hooking up to go.